Add newlines to fix tests on linux
The expected docker-config.yml did not contain a newline at the end
of the file.  This caused the test to fail on linux machines,
because the generated docker-config.yml will indeed have a newline
at the end of the file.

generateTest.js doesn't technically need a newline, but it's best practice.