Update CONTRIBUTING.md
This is an attempt at copying and adapting the contributing guide from ember-cli
to provide a little more structure and guidance for people looking to get involved with SANE.

Attempted to make the tone of the document a little more friendly. One way of doing this was
by blurring the line between 'you' and 'us' to make it clear that it is a team effort.
Also re-wrote the opening paragraph to be more inclusive and welcoming.

Removed the set of strict requirements on YUIDocs, since they were a little intimidating
and at this point it seems more important to just start using it.

Provided more details about issue reporting, and the benefits of documenting new features.