Allow SANE to run without global ember-cli
Uses npm-which to try to resolve a binary for ember first in local node_modules,
and then in parent node_modules and the user's path, the same way that a native
npm `require()` works.

More tests would be good to have, intent is to use rewire to inject process.platform
and process.cmd, but for now rewire does not work for modules
which return a simple string or variable with a string value.

See https://github.com/jhnns/rewire/issues/51