Use SANE's ember if no global ember exists
Previously, the `ember` helper would use the ember-cli closest to the
process.cwd, and if it couldn't find one in the current project, it would
use a global ember-cli.

The problem is that upon first installation of `sane-cli`, unless `ember-cli`
was also installed, the helper could not resolve a version of ember unless `sane` was
invoked inside the sane project.  This meant that commands like `sane new .` and even
`sane --help` would fail with `Error: not found: ember`.

The changes here will cause the ember helper to use the ember-cli within node_modules of
the globally installed sane-cli, much like the code in `commands/new.js`.

With this commit in place, installing `ember-cli` globally should no longer be necessary.